[TEST]: Do some tests for evaluation, almost finish every test. [TODO]: Final tests with CGI and Siege. [TODO]: Refactor of ugly functions, if possible.

+++ src/client/RequestHandler.cpp
@@ -3,18 +3,20 @@
 #include <unistd.h>
 
 bool	isAllowedMethod(Request* req, Response* res) {
-	if (res->getAllowedMethods().find(req->getMethod()) == res->getAllowedMethods().end())
-		return false;
-	return true;
+	std::set<std::string> allowedMethods = res->getAllowedMethods();
+	for (std::set<std::string>::iterator it = allowedMethods.begin(); it != allowedMethods.end(); it++) {
+		if (*it == req->getMethod())
+			return true;
+	}
+	return false;
 }
 
-void	handleRequest(Request* request, Http* http, Response* res, bool locationExists, int& statusCode, Upload *upload) {
-	(void)http;
+void	handleRequest(Request* request, Response* res, bool locationExists, int& statusCode, Upload *upload) {
 	if (request->getMethod() == "GET" && isAllowedMethod(request, res)) {
 		handleGet(request, res, locationExists, statusCode);
 	} else if (request->getMethod() == "POST" && isAllowedMethod(request, res)) {
 		handlePost(request, res, statusCode, upload);
-	} else if (request->getMethod() == "DELETE") {
+	} else if (request->getMethod() == "DELETE" && isAllowedMethod(request, res)) {
 		handleDelete(request, res, statusCode);
 	} else {
 		statusCode = 405;
@@ -22,15 +24,24 @@ void	handleRequest(Request* request, Http* http, Response* res, bool locationExi
 	}
 }
 
+bool	indexExists(const std::string& cwd, const std::string& root, const std::string& index) {
+	if (index.empty())
+		return false;
+	std::string path = cwd + root + "/" + index;
+	if (access(path.c_str(), F_OK) == 0)
+		return true;
+	return false;
+}
+
 void	handleGet(Request* req, Response* res, bool locationExists, int& statusCode) {
-	std::string cwd = getCurrentDir();
+	statusCode = 200;
+
+	std::string	cwd = getCurrentDir();
 	if (cwd.empty()) {
 		statusCode = 500;
 		res->setResponse(generateResponse(req, res));
 		return;
 	}
-
-	statusCode = 200;
 
 	if (req->getUrlPath().find("/favicon.ico") != std::string::npos) {
 		std::string path;
@@ -43,6 +54,7 @@ void	handleGet(Request* req, Response* res, bool locationExists, int& statusCode
 		return;
 	}
 
+	// REDIRECTION //
 	if (req->getUrlPath() == res->getRewriteToReplace()) {
 		if (res->getRewriteFlag() == "redirect")
 			statusCode = 307;
@@ -54,7 +66,8 @@ void	handleGet(Request* req, Response* res, bool locationExists, int& statusCode
 
 	if (isADirectory(req->getUrlPath(), res->getRoot())) {
 		// handle directory if location and index exist,
-		if (locationExists && locationMatches(req->getUrlPath(), res->getLocationPath()) && !res->getIndex().empty()) {
+		if (locationExists && locationMatches(req->getUrlPath(), res->getLocationPath())
+			&& indexExists(cwd, res->getRoot(), res->getIndex())) {
 			res->setPathForHtml(cwd + res->getRoot() + "/" + res->getIndex());
 		} // handle directory if location and autoindex exist
 		else if (locationExists && locationMatches(req->getUrlPath(), res->getLocationPath()) && res->getAutoindex() == true) {
@@ -65,7 +78,7 @@ void	handleGet(Request* req, Response* res, bool locationExists, int& statusCode
 			res->setResponse(generateResponse(req, res));
 			return;
 		} else {
-			if (!res->getIndex().empty()) {
+			if (indexExists(cwd, res->getRoot(), res->getIndex())) {
 				res->setPathForHtml(cwd + res->getRoot() + "/" + res->getIndex());
 			} else if (res->getAutoindex() == true) {
 				res->setResponse(generateDirectoryListing(req->getUrlPath(), res->getRoot(), res));
